Functions moved from views to urls, current weather and forecast as Tallinn on index.html
