+++ weatherapp/sweatherapp/templates/weather_result.html
@@ -12,38 +12,82 @@
             font-family: 'Arial', sans-serif;
             background-color: #f5f5f5;
             color: #333;
+            display: flex;
+            flex-direction: column;
+            align-items: center;
         }
 
         h1 {
             color: #007bff;
         }
 
-        /* Add more styles as needed */
+        .container {
+            display: flex;
+            justify-content: space-around;
+            width: 80%;
+            margin-top: 20px;
+        }
+
+        .current-weather,
+        .forecast {
+            width: 45%;
+            padding: 20px;
+            border: 1px solid #ccc;
+            margin-bottom: 20px;
+            box-sizing: border-box;
+        }
+
+        p {
+            margin-bottom: 10px;
+        }
+
+        ul {
+            list-style-type: none;
+            padding: 0;
+        }
+
+        li {
+            margin-bottom: 20px;
+        }
+
+        strong {
+            color: #007bff;
+        }
     </style>
 </head>
 <body>
 <h1>Current weather in {{ current_weather.city_name }}</h1>
 
-<p>Temperature: {{ current_weather.temperature|floatformat:0 }} °{{ temperature_unit }}</p>
-<p>Feels Like: {{ current_weather.feels_like|floatformat:0 }} °{{ temperature_unit }}</p>
-<p>Weather Description: {{ current_weather.weather_description }}</p>
-<p>Humidity: {{ current_weather.humidity }} %</p>
+<div class="container">
+    <div class="current-weather">
+        <p>Temperature: {{ current_weather.temperature|floatformat:0 }} °{{ temperature_unit }}</p>
+        <p>Feels Like: {{ current_weather.feels_like|floatformat:0 }} °{{ temperature_unit }}</p>
+        <p>Weather Description: {{ current_weather.weather_description }}</p>
+        <p>Humidity: {{ current_weather.humidity }} %</p>
+    </div>
 
-{% if forecast_data %}
-    <h2>Forecast for the next 5 days</h2>
-    <ul>
-        {% for forecast in forecast_data %}
-            <li>
-                <strong>{{ forecast.date_time|date:"Y-m-d H:i:s" }}</strong>
-                <br>
-                Temperature: {{ forecast.temperature|floatformat:0 }} °{{ temperature_unit }}
-                <br>
-                Feels Like: {{ forecast.feels_like|floatformat:0 }} °{{ temperature_unit }}
-                <br>
-                Weather Description: {{ forecast.weather_description }}
-                <br>
-                Humidity: {{ forecast.humidity }} %
-            </li>
-        {% endfor %}
-    </ul>
-{% endif %}
+    {% if forecast_data %}
+        <div class="forecast">
+            <h2>Forecast for the next 5 days</h2>
+            <ul>
+                {% for forecast in forecast_data %}
+                    {% if forloop.counter <= 5 %}
+                        <li>
+                            <strong>{{ forecast.date_time|date:"Y-m-d H:i:s" }}</strong>
+                            <br>
+                            Temperature: {{ forecast.temperature|floatformat:0 }} °{{ temperature_unit }}
+                            <br>
+                            Feels Like: {{ forecast.feels_like|floatformat:0 }} °{{ temperature_unit }}
+                            <br>
+                            Weather Description: {{ forecast.weather_description }}
+                            <br>
+                            Humidity: {{ forecast.humidity }} %
+                        </li>
+                    {% endif %}
+                {% endfor %}
+            </ul>
+        </div>
+    {% endif %}
+</div>
+</body>
+</html>
